Move site object to a json data file
