+++ src/data/site.json
@@ -1,0 +1,7 @@
+{
+  "title": "Fluid Type Scale Calculator",
+  "metaTitle": "Fluid Type Scale - Generate responsive font-size variables",
+  "description": "Generate font size variables for a fluid type scale with CSS clamp. Grab the output CSS and drop it into any existing design system.",
+  "keywords": ["fluid type scale", "type scale", "CSS clamp"],
+  "url": "https://www.fluid-type-scale.com"
+}
