+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import FluidTypeScaleCalculator from '../components/FluidTypeScaleCalculator/Flu
 import HeroBanner from '../components/HeroBanner/HeroBanner';
 import Info from '../components/Info/Info';
 import Layout from '../components/Layout/Layout';
-import { site } from '../constants';
+import site from '../data/site.json';
 import type { WithFonts } from '../types';
 import { getGoogleFontFamilies } from '../utils';
